feat(inventory): enhance stock adjustment dialog with new form fields and logic
- Added `StockAdjustFormFields` component to encapsulate form fields for stock adjustments.
- Introduced `useStockAdjustDialog` hook to manage dialog state and operations.
- Updated `StockAdjustDialog` to utilize the new form fields and handle operations more efficiently.
- Integrated inventory-related types and operations for better type safety and functionality.
